refactor(Our menu): extract GetMenuDataProps and ValidatePaginationParamsProps into types/menu.ts
