+++ src/types/menu.ts
@@ -1,3 +1,4 @@
+import { RecipeProps } from "./recipe";
 interface MenuParamsProps {
   params: {
     preference: string;
@@ -8,17 +9,44 @@ interface MenuParamsProps {
   };
 }
 
-interface MenuTabsProps{
-  page: number,
-  limit: number,
-  active: string
+interface MenuTabsProps {
+  page: number;
+  limit: number;
+  active: string;
 }
 
 interface PaginationProps {
-  currentPage: number,
-  totalPages: number,
-  preference: string,
-  limit: number
+  currentPage: number;
+  totalPages: number;
+  preference: string;
+  limit: number;
 }
 
-export type { MenuParamsProps, MenuTabsProps, PaginationProps }
+type GetMenuDataProps =
+  | { previewOnly: true; totalItems: number }
+  | {
+      previewOnly?: false;
+      recipes: RecipeProps[];
+      pagination: {
+        currentPage: number;
+        totalPages: number;
+        next: number;
+        prev: number;
+        pages: number[];
+      };
+    };
+
+interface ValidatePaginationParamsProps {
+  page: number;
+  limit: number;
+  totalItems: number;
+  basePath: string;
+}
+
+export type {
+  MenuParamsProps,
+  MenuTabsProps,
+  PaginationProps,
+  GetMenuDataProps,
+  ValidatePaginationParamsProps
+};
